上位机/waist_ui/requirements.txt
PySide6
PyQt-Fluent-Widgets
paho-mqtt
python-dotenv
